[af-62] rename: dto 디렉토리에 request, resonse 하위 디렉토리 생성
