+++ src/main/java/com/drunkenlion/alcoholfriday/domain/cart/application/CartService.java
@@ -1,9 +1,9 @@
 package com.drunkenlion.alcoholfriday.domain.cart.application;
 
-import com.drunkenlion.alcoholfriday.domain.cart.dto.CartDetailResponse;
-import com.drunkenlion.alcoholfriday.domain.cart.dto.CartRequest;
-import com.drunkenlion.alcoholfriday.domain.cart.dto.CartResponse;
-import com.drunkenlion.alcoholfriday.domain.cart.dto.DeleteCartRequest;
+import com.drunkenlion.alcoholfriday.domain.cart.dto.response.CartDetailResponse;
+import com.drunkenlion.alcoholfriday.domain.cart.dto.request.CartRequest;
+import com.drunkenlion.alcoholfriday.domain.cart.dto.response.CartResponse;
+import com.drunkenlion.alcoholfriday.domain.cart.dto.request.DeleteCartRequest;
 import com.drunkenlion.alcoholfriday.domain.cart.entity.Cart;
 import com.drunkenlion.alcoholfriday.domain.member.entity.Member;
