+++ src/test/java/com/drunkenlion/alcoholfriday/domain/cart/application/CartServiceTest.java
@@ -3,10 +3,10 @@ package com.drunkenlion.alcoholfriday.domain.cart.application;
 import com.drunkenlion.alcoholfriday.domain.auth.enumerated.ProviderType;
 import com.drunkenlion.alcoholfriday.domain.cart.dao.CartDetailRepository;
 import com.drunkenlion.alcoholfriday.domain.cart.dao.CartRepository;
-import com.drunkenlion.alcoholfriday.domain.cart.dto.CartDetailResponse;
-import com.drunkenlion.alcoholfriday.domain.cart.dto.CartRequest;
-import com.drunkenlion.alcoholfriday.domain.cart.dto.CartResponse;
-import com.drunkenlion.alcoholfriday.domain.cart.dto.DeleteCartRequest;
+import com.drunkenlion.alcoholfriday.domain.cart.dto.response.CartDetailResponse;
+import com.drunkenlion.alcoholfriday.domain.cart.dto.request.CartRequest;
+import com.drunkenlion.alcoholfriday.domain.cart.dto.response.CartResponse;
+import com.drunkenlion.alcoholfriday.domain.cart.dto.request.DeleteCartRequest;
 import com.drunkenlion.alcoholfriday.domain.cart.entity.Cart;
 import com.drunkenlion.alcoholfriday.domain.cart.entity.CartDetail;
 import com.drunkenlion.alcoholfriday.domain.category.entity.Category;
